+++ AST.h
@@ -4,7 +4,7 @@
 /* unary-and-binary tree structure */
 struct _tree {
   char car;                    	/* char for arithmetic operation */
-  int val;			/* int  for value */
+  double val;			/* int  for value */
   struct _tree* left;    	/* used for unary node but NULL if leaf */
   struct _tree* right;   	/* NULL if unary node or leaf*/
 };
@@ -18,7 +18,7 @@ AST newBinaryAST(char car, AST left, AST right);
 AST newUnaryAST(char car, AST son);
 
 /* create an AST leaf from a value */
-AST newLeafAST(int val);
+AST newLeafAST(double val);
 
 /* delete an AST */
 void freeAST(AST t);
